Decouple ExportCodeInput from CodeInput
